feat: song request scaffold
